Unethical Aspects of Image Classification
This project was a collaborative paper on Image Classification.

Image classification has played an important role in everyday uses and the growth of data
science. According to the case study featured on Science Direct, Image Classification refers to
the process of computer vision being able to classify an image to its visual content. The use of
algorithms to classify is used in everyday life from experimentation to visual tagging on social
media. The algorithms used for image classification are able to be seen as both positive and
negative within the data science community.
Our paper is presented to the reader as follows; the first section will showcase image
classification models used within multiple industry fields and how it can benefit society. The
second section will feature the unethical aspects of image classification. Our final section will
summarize what we discovered and discuss why our group believes this is the importance to the
data science community
Sources to cite in these sections for ACM
https://www.sciencedirect.com/topics/computer-science/image-classification